Refactor matrix switch component into nested component
