+++ components/matrix_display/display.py
@@ -10,7 +10,7 @@ from esphome.const import (
     CONF_WIDTH,
 )
 
-AUTO_LOAD = ["switch", "matrix_display_switch", "number", "matrix_display_brightness"]
+AUTO_LOAD = ["number", "matrix_display_brightness"]
 MATRIX_ID = "matrix_id"
 CHAIN_LENGTH = "chain_length"
 BRIGHTNESS = "brightness"
